Modifies genomic regions by centering and then expanding them

center_expand_regions is an optional step that re-defines the genomic regions by expanding them from their center. The center information has to be stored in the input data column center, while the information for the expansion can either be user provided or input data derived. The accepted input is a data frame created from prepare_input_regions. Please see prepare_input_regions for more details.

center_expand_regions(
  data,
  center_by = "center_column",
  expand_by = NULL,
  show_messages = TRUE
)

Arguments

data

PeakCombiner data frame structure with required columns named chrom, start, end, name, score, strand, center, sample_name. Additional columns will be maintained.

center_by

Allowed values are 'center_column' (default) or 'midpoint'.

'center_column' uses the value stored in the column center to center.

'midpoint' replaces the value stored in the column center with the mathematical mean of each genomic region (e.g., round(end - start / 2)), which is then used.

expand_by

Allowed values a numeric vector of length 1 or 2, or 'NULL' (default).

The value from the numeric vector of length 1 is expanded in both directions from center to define the genomic region. Thus, the size of the resulting genomic region is 2x the provided value + 1 (for the center coordinate).

The value of the numeric vector of length 2 subtracts the first value from the center and adds the second value to the center to define the genomic region. Thus, the size of the genomic regions is the sum of the first value + the second value + 1 (for the center coordinate).

'NULL' allows for data-driven definition of the expand_by value. It calculates the median genomic region size of the input data and uses this value like a length 1 numeric vector for expansion.

show_messages

Logical value of TRUE (default) or FALSE. Defines if info messages are displayed or not.

Value

A tibble with the columns chrom, start, end, name, score, strand, center, sample_name. The definitions of these columns are described in full in the prepare_input_regions Details. Use as input for functions filter_regions() and combine_regions().

Details

This is an optional function that resizes the genomic regions based on the input peakCombiner standard data frame and the options you select. An expected input data foam contains the following columns with the names: chrom, start, end, name, score, strand, center, sample_name. Such a data frame is created by the script prepare_input_regions. This step is useful if you want all of your peaks to be the same size for your downstream analyses. In addition, if you want to use the "summit" information, normally obtained by some peak callers (e.g., Macs2), this function allows you to automatically center your regions of interest on these summits. This enables you to capture information about the most important region within a genomic region (e.g., TF-binding site or highest peak) and put that region in the center of your downstream analyses (e.g., applicable to motif-finding or "heatmaps" summarizing multiple genomic regions).

There are two concepts that are relevant for center_expand_regions: how to define the center, and how much to expand from the center.

How to define the center?

When you prepared your input regions, it is recommended to use the function prepare_input_regions provided by this package. This pre- populated the center column with the absolute genomic coordinate of the center of the peak region. You can either choose to define the center by using pre-defined summit information (e.g., obtained from a peak caller like MACS2) or re-compute the arithmetic mean and save that value in the column center. (For details see the help for prepare_input_regions).

How much to expand from the center

You can choose to expand the genomic region from the center either symmetrically or asymmetrically (different lengths before and after the center position).

In the symmetrical case, if you want to choose the size of your genomic region based on the input data, this function can also calculate the median peak size across all of your genomic regions and use that value (expand_by = NULL). Alternatively, the user is free to provide a numeric vector to define the expansion. A numeric vector with one value is used to symmetrically expand, while a vector with two values allows to expand asymmetrically.

Examples

# Load in and prepare a an accepted tibble
utils::data(syn_data_bed)

# Prepare input data
data_prepared <- prepare_input_regions(
  data = syn_data_bed,
  show_messages = TRUE
)
#> ℹ Provide input `data` is a pre-loaded <data.frame> with the required column
#>   names.
#> → Start preparing data.
#> ℹ Required columns will be added if missing.
#> ! Column 'score' does not exist in `data_prepared`.
#> → Column 'score' is added and filled with "0".
#> ! Column 'strand' does not exist in `data_prepared`.
#> → Column 'strand' is added and filled with ".".
#> ! Column 'summit' does not exist in `data_prepared`.
#> → As no input column 'summit' is found, the output column 'center' has to be
#>   filled with arithmetic center of peak.
#> → Checking <class> and "values" of all columns.
#> ✔ Structure of data was successfully checked to be an accepted input.
#>   
#> → Checking whether duplicated regions exist and need to be collapsed.
#> → Checked whether duplicated regions exist and need to be collapsed.
#> ✔ Duplicated regions identified and collapsed to unique chrom, start, and end
#>   for each sample by strongest score value.
#>   
#> ✔ Preparation of data finished successfully.
# Run center and expand
data_center_expand <- center_expand_regions(
  data = data_prepared,
  center_by = "center_column",
  expand_by = NULL,
  show_messages = TRUE
)
#> ℹ Input value for `expand_by` is "NULL". Median of all input genomic regions is
#>   calculated and returned for expansion.
#> ✔ `expand_by` was calculated from the input data and set to "350".
#> ℹ Genomic regions will be expanded by 350bp in both direction.
#>   
#> → Checking <class> and "values" of all columns.
#> ✔ Structure of data was successfully checked to be an accepted input.
#>   
#> → Genomic regions will be centered and expanded.
#>   
#> → Starting with expanding genomic regions from the column center.
#> ✔ Genomic regions were successfully centered and expanded.
#>   
#> ℹ Some newly-defined genomic regions have a start coordinate below "1".
#> → Values of name for these site: "control_rep2|1".
#> ✔ These genomic regions were truncated to get start coordinate "1".
#> ✔ Genomic regions were successfully centered and expanded.
#>   

data_center_expand
#> # A tibble: 52 × 8
#>    chrom  start   end name            score strand center sample_name 
#>    <chr>  <dbl> <dbl> <chr>           <dbl> <chr>   <dbl> <chr>       
#>  1 Chr2    100.  800. control_rep1|9      0 .        450. control_rep1
#>  2 chr1    300. 1000. control_rep1|1      0 .        650. control_rep1
#>  3 chr1   3600. 4300. control_rep1|2      0 .       3950. control_rep1
#>  4 chr1   4650. 5350. control_rep1|3      0 .       5000. control_rep1
#>  5 chr1   5500. 6200. control_rep1|5      0 .       5850. control_rep1
#>  6 chr1   6700. 7400. control_rep1|6      0 .       7050. control_rep1
#>  7 chr10   300. 1000. control_rep1|7      0 .        650. control_rep1
#>  8 chr2    300. 1000. control_rep1|8      0 .        650. control_rep1
#>  9 chr4 2  300. 1000. control_rep1|10     0 .        650. control_rep1
#> 10 chr4-2  400. 1100. control_rep1|11     0 .        750. control_rep1
#> # ℹ 42 more rows

# You can choose to use the midpoint and predefined values to expand

data_center_expand <- center_expand_regions(
  data = data_prepared,
  center_by = "midpoint",
  expand_by = c(100, 600),
  show_messages = FALSE
)

data_center_expand
#> # A tibble: 52 × 8
#>    chrom  start   end name            score strand center sample_name 
#>    <chr>  <dbl> <dbl> <chr>           <dbl> <chr>   <dbl> <chr>       
#>  1 Chr2     350  1050 control_rep1|9      0 .         450 control_rep1
#>  2 chr1     550  1250 control_rep1|1      0 .         650 control_rep1
#>  3 chr1    3850  4550 control_rep1|2      0 .        3950 control_rep1
#>  4 chr1    4900  5600 control_rep1|3      0 .        5000 control_rep1
#>  5 chr1    5750  6450 control_rep1|5      0 .        5850 control_rep1
#>  6 chr1    6950  7650 control_rep1|6      0 .        7050 control_rep1
#>  7 chr10    550  1250 control_rep1|7      0 .         650 control_rep1
#>  8 chr2     550  1250 control_rep1|8      0 .         650 control_rep1
#>  9 chr4 2   550  1250 control_rep1|10     0 .         650 control_rep1
#> 10 chr4-2   650  1350 control_rep1|11     0 .         750 control_rep1
#> # ℹ 42 more rows