Combine overlapping genomic regions from different samples to create a single set of consensus genomic regions

combine_regions is the main function of this package and combines overlapping genomic regions from different samples to create a single set of consensus genomic regions.

The accepted input is the PeakCombiner data frame is created from the function prepare_input_regions and has optionally already been centered and expanded and / or filtered using center_expand_regions and filter_regions, respectively. Please see prepare_input_regions for more details.

combine_regions(
  data,
  found_in_samples = 2,
  combined_center = "nearest",
  annotate_with_input_names = FALSE,
  combined_sample_name = NULL,
  show_messages = TRUE
)

Arguments

data

PeakCombiner data frame structure with required columns named chrom, start, end, name, score, strand, center, sample_name. Additional columns will be dropped

found_in_samples

Only include genomic regions that are found in at least found_in_samples number of samples. If found_in_samples is a fraction between 0 and 1, then only include genomic regions that ar found in at least found_in_samples fraction of samples. Default value is 2.

combined_center

Defines how the column 'center' will be populated for each genomic region in the output data. Allowed options are * middle - the mathematical center of the new region * strongest - the 'center' of the input region that has the the highest 'score' of all overlapping input regions * nearest - the 'center' of the input region that is closest to mean of the 'center's of all overlapping input regions (default)

annotate_with_input_names

TRUE / FALSE (default). If TRUE, a new column named 'input_names' is created in the output data that is populated for each combined genomic region with the 'name's of all contributing input regions. If the column 'input_names' already exists, it will be overwritten.

combined_sample_name

Optionally defines how the column 'sample_name' is populated for the output data. If not used, then the default is to simply concatenate all input sample_names into a single comma-separated string

show_messages

Logical value of TRUE (default) or FALSE. Defines if info messages are displayed or not.

Value

A tibble with the columns chrom, start, end, name, score, strand, center, sample_name, and optionally input_names. The definitions of these columns are described in full in the Details below. Use as input for functions center_expand_regions and filter_regions.

Details

combine_regions creates a set of consensus genomic regions by combining overlapping genomic regions from different samples. The general steps within this function are:

Identify overlapping genomic regions from the input samples

Retain overlapping genomic regions that are found in at least found_in_samples samples. In this way, you can remove rare or sample-specific regions

Note that overlapping genomic regions must contain at least one 'center' from its input sample regions to be considered a valid genomic region.

As you can use the output data from this step again (e.g., to center and expand the new set of consensus regions), we must define the 'center', 'score', 'sample_name', and 'name' values for the new genomic regions. We do this as follows:

'center' is defined by the combined_center parameter, which has three options. * middle - the mathematical center of the new region * strongest - the 'center' of the input region that has the the highest 'score' of all overlapping input regions * nearest - the 'center' of the input region that is closest to mean of the 'center's of all overlapping input regions (default)

'score' is the score of the genomic region from the sample whose 'center's was used, or the mean of the 'score's if middle was selected for the combined_center parameter

'sample_name' can be user defined (combined_sample_name) or is a concatenated string of all input 'sample_names' (default).

'name' is created by combining 'sample_name' and row number to create a unique identifier for each newly created genomic region.

Note, the output data.frame columns sample_name, name and score will be updated.

Examples

# Load in and prepare a an accepted tibble
utils::data(syn_data_bed)

data_prepared <- prepare_input_regions(
  data = syn_data_bed,
  show_messages = FALSE
)

# Lets combine the input data by defining all potential option
combine_regions(
  data = data_prepared,
  found_in_samples = 2,
  combined_center = "nearest",
  annotate_with_input_names = TRUE,
  combined_sample_name = "consensus",
  show_messages = TRUE
)
#> → Checking <class> and "values" of all columns.
#> ✔ Structure of data was successfully checked to be an accepted input.
#>   
#> → Start with the disjoining and filtering genomic regions.
#> ✔ Disjoin and filter by `found_in_samples` of genomic regions successfully
#>   finished.
#>   
#> → Start with combining remaining genomic regions.
#> ✔ Combining remaining genomic regions was successfully finished.
#>   
#> → Start with identification of overlaps between the original summit and
#>   remaining genomic regions.
#> ℹ Remaining regions without overlap will be removed.
#> Joining with `by = join_by(ranking)`
#> ✔ Retained genomic regions with input data summit overlap was successfully
#>   finished.
#>   
#> → Information from input center will be added to output data frame.
#> ℹ Argument `combined_center` was defined as "nearest".
#> ℹ The mean of all input centers is calculated and the nearest input center is
#>   used
#> → Center information in center and score are added to the output data frame.
#> ✔ Output data frame columns center and score were successfully populated.
#>   
#> → The value "consensus" for column sample_name was provided.
#> ℹ Column sample_name is filled with provided value consensus.
#> ℹ Column name is created as unique identifier for each row containing
#>   sample_name and the row number.
#> ✔ The columns sample_name and name were successfully populated.
#>   
#> ℹ Argument `annotate_with_input_names` was set to "TRUE".
#> → Column input_names is added to output data frame.
#> ✔ Additional column input_names was successfully populated.
#>   
#> ✔ All required columns in output data frame were successfully populated.
#>   
#> ✔ Genomic regions were successfully combined.
#>   
#> # A tibble: 10 × 9
#>    chrom  start   end name         score strand center sample_name input_names  
#>    <chr>  <dbl> <dbl> <chr>        <dbl> <chr>   <dbl> <chr>       <chr>        
#>  1 chr1     101  1000 consensus|1      0 .        550  consensus   control_rep1…
#>  2 chr1    2501  3200 consensus|2      0 .       2900. consensus   treatment_re…
#>  3 chr1    3501  4400 consensus|3      0 .       3950. consensus   control_rep1…
#>  4 chr1    4501  5300 consensus|4      0 .       4900. consensus   control_rep1…
#>  5 chr1    5701  6300 consensus|5      0 .       6050. consensus   control_rep1…
#>  6 chr1    6701  7400 consensus|6      0 .       7050. consensus   control_rep1…
#>  7 chr2     101  1000 consensus|7      0 .        550. consensus   control_rep1…
#>  8 chr10    101  1000 consensus|8      0 .        550. consensus   control_rep1…
#>  9 chr4 2   301  1000 consensus|9      0 .        650. consensus   control_rep1…
#> 10 chr4-2   401   900 consensus|10     0 .        550. consensus   control_rep1…