Update the center and score information

Helper function for main function combine_regions. Requires in memory data frame in the standard accepted format for the peakCombiner package. For details see the details for combine_regions.

cr_add_summit(
  data,
  input,
  combined_center = "nearest",
  annotate_with_input_names = FALSE,
  combined_sample_name = NULL
)

Arguments

data

PeakCombiner data frame structure with required columns named chrom, start, end, name, score, strand, center, sample_name. Additional columns will be dropped

input

The original input file from combine_regions to extract center information

combined_center

Defines how the column 'center' will be populated for each genomic region in the output data. Allowed options are * middle - the mathematical center of the new region * strongest - the 'center' of the input region that has the the highest 'score' of all overlapping input regions * nearest - the 'center' of the input region that is closest to mean of the 'center's of all overlapping input regions (default)

annotate_with_input_names

TRUE / FALSE (default). If TRUE, a new column named 'input_names' is created in the output data that is populated for each combined genomic region with the 'name's of all contributing input regions. If the column 'input_names' already exists, it will be overwritten.

combined_sample_name

Optionally defines how the column 'sample_name' is populated for the output data. If not used, then the default is to simply concatenate all input sample_names into a single comma-separated string

Value

A tibble with the following columns: chrom, start, end, name, score, strand, center, sample_name.

Details

As you can use the output data from this step again (e.g., to center and expand the new set of consensus regions), we must define the 'center', 'score', 'sample_name', and 'name' values for the new genomic regions. We do this as follows:

'center' is defined by the combined_center parameter, which has three options. * middle - the mathematical center of the new region * strongest - the 'center' of the input region that has the the highest 'score' of all overlapping input regions * nearest - the 'center' of the input region that is closest to mean of the 'center's of all overlapping input regions (default)

'score' is the score of the genomic region from the sample whose 'center's was used, or the mean of the 'score's if middle was selected for the combined_center parameter

'sample_name' is a concatenated string of all input sample_names

In addition, the output data.frame columns sample_name, name and score will be updated.