Apply user-defined filtering options to genomic regions.

filterRegions is an optional step that allows inclusion or exclusion of genomic regions based on 4 different criteria:

Include regions by their chromosome names (optional).

Exclude blacklisted regions (optional).

Include regions above a given score (optional).

Include top n regions per sample, ranked from highest to lowest score (optional).

The accepted input is the PeakCombiner data frame is created from the function prepareInputRegions. Please see prepareInputRegions for more details.

The filterRegions can be used multiple times on the same data set, which allows a user to step-wise optimize selection criteria of regions of interest.

filterRegions(
  data,
  includeByChromosomeName = NULL,
  excludeByBlacklist = NULL,
  includeAboveScoreCutoff = NULL,
  includeTopNScoring = NULL,
  outputFormat = "GenomicRanges",
  showMessages = TRUE
)

Arguments

data

PeakCombiner data frame structure with required columns named chrom, start, end, name, score, strand, center, sample_name. Additional columns will be maintained.

includeByChromosomeName

'NULL' (default) - No chromosome name filtering will be done. * Character vector that contains chromosomes names to be retained.

excludeByBlacklist

'NULL' (default) - No blacklist filtering will be done. * GenomicRanges object (default setup) or data frame/tibble with columns chrom, start, and end.

includeAboveScoreCutoff

'NULL' (default) - No score filtering will be done. * Single numeric value that defines the score threshold above which all genomic regions will be retained. This results in variable number of sites per sample.

includeTopNScoring

'NULL' (default) - No score filtering will be done. * Single numeric value representing the number of genomic regions per sample to be retained. The genomic regions are selected from highest to lowest score, and if includeTopNScoring > number of regions, then no filtering is done.

outputFormat

Character value to define format of output object. Accepted values are "GenomicRanges" (default), "tibble" or "data.frame".

showMessages

Logical value of TRUE (default) or FALSE. Defines if info messages are displayed or not.

Value

A tibble with the columns chrom, start, end, name, score, strand, center, sample_name. The definitions of these columns are described in full in the prepareInputRegions Details. Use as input for functions centerExpandRegions and combineRegions.

Details

This is an optional step which enables commonly-needed filtering steps to focus in on the key genomic regions of interest. This can be useful when there are many genomic regions identified in your peak-caller or input BED files.

filterRegions can be used multiple times on the same data set, allowing a user to select regions of interest using a step-wise optimization approach.

includeByChromosomeName - Retains only chromosomes that are in the provided vector. By not including mitochondrial, sex, or non-classical chromosomes, genomic regions found on these chromosomes can be removed. If set to 'NULL' (default), this step will be skipped (optional).

excludeByBlacklist - A GenomicRanges file, dataframe or tibble can be provided listing the genomic regions to remove (having chrom ( seqnames for GenomicRanges) , start, and end column names). If set to 'NULL' (default), this step will be skipped (optional). Please note that if there are not matching entries in the 'chrom' columns of input and blacklist, an information message is displayed. This can happend and does not cause any problems with the script.

includeAboveScoreCutoff - Single numeric value that defines the score threshold above which all genomic regions will be retained. The score column in the peakCombiner input data should be non-zero for this parameter to be used. It is populated by prepareInputRegions, and by default takes the value of -log10(FDR) if possible (e.g., using a .narrowPeak file from MACS2 as input). Importantly, applying this filter retains a variable number of genomic regions per sample, all having a score greater than the includeAboveScoreCutoff parameter. If set to 'NULL' (default), this step will be skipped (optional).

includeTopNScoring - Single numeric value that defines how many of the top scoring genomic regions (using the column score) are retained. All other genomic regions are discarded. Importantly, applying this filter retains includeTopNScoring regions per sample, which means that the minimum enrichment levels may vary between samples. Note that if multiple genomic regions have the same score cutoff value, then all of those genomic regions are included. In this case, the number of resulting regions retained may be a bit higher than the input parameter. If set to 'NULL' (default), this step will be skipped (optional).

Examples

# Load in and prepare a an accepted tibble
utils::data(syn_data_bed)

data_prepared <- prepareInputRegions(
  data = syn_data_bed,
  outputFormat = "tibble",
  showMessages = TRUE
)
#> ℹ Argument `outputFormat` is set to "tibble".
#> ℹ Argument `startsAreBased` is 1.
#> ℹ Argument `starts.in.df.are.0based` is FALSE.
#> ℹ Provide input `data` is a pre-loaded <data.frame> with the required column
#>   names.
#> → Start preparing data.
#> ℹ Required columns will be added if missing.
#> ! Column 'score' does not exist in `data_prepared`.
#> → Column 'score' is added and filled with "0".
#> ! Column 'strand' does not exist in `data_prepared`.
#> → Column 'strand' is added and filled with ".".
#> ! Column 'summit' does not exist in `data_prepared`.
#> → As no input column 'summit' is found, the output column 'center' has to be
#>   filled with arithmetic center of peak.
#> → Checking <class> and "values" of all columns.
#> ✔ Structure of data was successfully checked to be an accepted input.
#>   
#> → Checking whether duplicated regions exist and need to be collapsed.
#> → Checked whether duplicated regions exist and need to be collapsed.
#> ✔ Duplicated regions identified and collapsed to unique chrom, start, and end
#>   for each sample by strongest score value.
#>   
#> ✔ Preparation of data finished successfully.
#> ℹ Output format is set to "tibble".

# Here use options for all four filtering methods.

filterRegions(
  data = data_prepared,
  includeByChromosomeName = c("chr1", "chr2", "chr4"),
  excludeByBlacklist = NULL,
  includeAboveScoreCutoff = 10,
  includeTopNScoring = 100,
  outputFormat = "tibble",
  showMessages = TRUE
)
#> → Checking <class> and "values" of all columns.
#> ✔ Structure of data was successfully checked to be an accepted input.
#>   
#> ℹ Provide input `data` is a <data.frame> with three or four columns and paths
#>   to existing files.
#> → Start loading and preparing data.
#> ℹ Argument `outputFormat` is set to "tibble".
#> → The argument `includeByChromosomeName` is a class <character> of length 3 and
#>   will be used to retain matchhing chromsome names in chrom.
#> ! Input for `includeByChromosomeName` contains values not found in the input
#>   data.
#> ℹ The following chromosome name you entered is not used: "chr4".
#> → Is "includeByChromosomeName" correctly defined?
#> ✔ Entries in chrom with the values "chr1" and "chr2" are retained.
#> ℹ The following 7 entries in column chrom from the input data were not
#>   retained: "Chr2", "chr10", "chr4 2", "chr4-2", "chr4|2", "chr42", and
#>   "chr4?2".
#> ✔ Input data was filtered to retain regions on defined chromosome.
#>   
#> ℹ The argument `excludeByBlacklist` is "NULL".
#> ✔ No filtering by blacklisted regions is done.
#>   
#> → Significance in score is filtered and all regions above 10 will be retained.
#> ℹ A total of 0 of 36 input regions are retained with value in score a above 10.
#> ✔ Input data was filtered to retain regions with a score above the defined
#>   threshold.
#>   
#> ℹ The argument `includeTopNScoring` extracted the the top 100 regions by score
#>   per sample (based on the values in sample_name).
#> → The top enriched 100 regions per sample will be retained.
#> ✔ Input data was filtered and the top 100 enriched regions per sample are
#>   retained.
#>   
#> ✔ Filtered dataset will be returned.
#> ℹ Output format is set to "tibble".
#> # A tibble: 0 × 8
#> # ℹ 8 variables: chrom <chr>, start <dbl>, end <dbl>, name <chr>, score <dbl>,
#> #   strand <lgl>, center <dbl>, sample_name <chr>