Apply user-defined filtering options to genomic regions.

filter_regions is an optional step that allows inclusion or exclusion of genomic regions based on 4 different criteria:

Include regions by their chromosome names (optional).

Exclude blacklisted regions (optional).

Include regions above a given score (optional).

Include top n regions per sample, ranked from highest to lowest score (optional).

The accepted input is the PeakCombiner data frame is created from the function prepare_input_regions. Please see prepare_input_regions for more details.

The filter_regions can be used multiple times on the same data set, which allows a user to step-wise optimize selection criteria of regions of interest.

filter_regions(
  data,
  include_by_chromosome_name = NULL,
  exclude_by_blacklist = NULL,
  include_above_score_cutoff = NULL,
  include_top_n_scoring = NULL,
  show_messages = TRUE
)

Arguments

data

PeakCombiner data frame structure with required columns named chrom, start, end, name, score, strand, center, sample_name. Additional columns will be maintained.

include_by_chromosome_name

'NULL' (default) - No chromosome name filtering will be done. * Character vector that contains chromosomes names to be retained.

exclude_by_blacklist

'NULL' (default) - No blacklist filtering will be done. * Single value for annotation of provided blacklists ('hg38' or 'mm10'), which uses ENCODE blacklist to filter. * Data frame or tibble with columns chrom, start, and end.

include_above_score_cutoff

'NULL' (default) - No score filtering will be done. * Single numeric value that defines the score threshold above which all genomic regions will be retained. This results in variable number of sites per sample.

include_top_n_scoring

'NULL' (default) - No score filtering will be done. * Single numeric value representing the number of genomic regions per sample to be retained. The genomic regions are selected from highest to lowest score, and if include_top_n_scoring > number of regions, then no filtering is done.

show_messages

Logical value of TRUE (default) or FALSE. Defines if info messages are displayed or not.

Value

A tibble with the columns chrom, start, end, name, score, strand, center, sample_name. The definitions of these columns are described in full in the prepare_input_regions Details. Use as input for functions center_expand_regions and combine_regions.

Details

This is an optional step which enables commonly-needed filtering steps to focus in on the key genomic regions of interest. This can be useful when there are many genomic regions identified in your peak-caller or input BED files.

filter_regions can be used multiple times on the same data set, allowing a user to select regions of interest using a step-wise optimization approach.

include_by_chromosome_name - Retains only chromosomes that are in the provided vector. By not including mitochondrial, sex, or non-classical chromosomes, genomic regions found on these chromosomes can be removed. If set to 'NULL' (default), this step will be skipped (optional).

exclude_by_blacklist - Remove ENCODE-annotated blacklisted regions for either human ( hg38 only) or mouse ( mm10 only). Alternatively, a data frame or tibble can be provided listing the genomic regions to remove (having chrom, start, and end column names). If set to 'NULL' (default), this step will be skipped (optional). Please note that if there are not matching entries in the 'chrom' columns of input and blacklist, an information message is displayed. This can happend und does not cause any problems with the script.

include_above_score_cutoff - Single numeric value that defines the score threshold above which all genomic regions will be retained. The score column in the peakCombiner input data should be non-zero for this parameter to be used. It is populated by prepare_input_regions, and by default takes the value of -log10(FDR) if possible (e.g., using a .narrowPeak file from MACS2 as input). Importantly, applying this filter retains a variable number of genomic regions per sample, all having a score greater than the include_above_score_cutoff parameter. If set to 'NULL' (default), this step will be skipped (optional).

include_top_n_scoring - Single numeric value that defines how many of the top scoring genomic regions (using the column score) are retained. All other genomic regions are discarded. Importantly, applying this filter retains include_top_n_scoring regions per sample, which means that the minimum enrichment levels may vary between samples. Note that if multiple genomic regions have the same score cutoff value, then all of those genomic regions are included. In this case, the number of resulting regions retained may be a bit higher than the input parameter. If set to 'NULL' (default), this step will be skipped (optional).

Examples

# Load in and prepare a an accepted tibble
input_data <- peakCombiner::syn_data_bed
input_data
#> # A tibble: 55 × 4
#>    chrom start   end sample_name   
#>    <chr> <dbl> <dbl> <chr>         
#>  1 chr1    200   900 treatment_rep1
#>  2 chr1      1   900 treatment_rep3
#>  3 chr1    101   300 control_rep2  
#>  4 chr1    301   900 control_rep2  
#>  5 chr1    201   900 treatment_rep1
#>  6 chr1    301   900 treatment_rep2
#>  7 chr1    301  1000 control_rep1  
#>  8 chr1    301  1100 control_rep3  
#>  9 chr1   1301  1600 treatment_rep3
#> 10 chr1   1901  2200 treatment_rep2
#> # ℹ 45 more rows

data_prepared <- prepare_input_regions(
  data = input_data,
  show_messages = TRUE
)
#> ℹ Provide input `data` is a pre-loaded <data.frame> with the required column
#>   names.
#> → Start preparing data.
#> ℹ Required columns will be added if missing.
#> ! Column 'score' does not exist in `data_prepared`.
#> → Column 'score' is added and filled with "0".
#> ! Column 'strand' does not exist in `data_prepared`.
#> → Column 'strand' is added and filled with ".".
#> ! Column 'summit' does not exist in `data_prepared`.
#> → As no input column 'summit' is found, the output column 'center' has to be
#>   filled with arithmetic center of peak.
#> → Checking <class> and "values" of all columns.
#> ✔ Structure of data was successfully checked to be an accepted input.
#>   
#> → Checking whether duplicated regions exist and need to be collapsed.
#> → Checked whether duplicated regions exist and need to be collapsed.
#> ✔ Duplicated regions identified and collapsed to unique chrom, start, and end
#>   for each sample by strongest score value.
#>   
#> ✔ Preparation of data finished successfully.

# Here use options for all four filtering methods.

filter_regions(
  data = data_prepared,
  include_by_chromosome_name = c("chr1", "chr2", "chr4"),
  exclude_by_blacklist = "hg38",
  include_above_score_cutoff = 10,
  include_top_n_scoring = 100,
  show_messages = TRUE
)
#> → Checking <class> and "values" of all columns.
#> ✔ Structure of data was successfully checked to be an accepted input.
#>   
#> → The argument `include_by_chromosome_name` is a class <character> of length 3
#>   and will be used to retain matchhing chromsome names in chrom.
#> ! Input for `include_by_chromosome_name` contains values not found in the input
#>   data.
#> ℹ The following chromosome name you entered is not used: "chr4".
#> → Is "include_by_chromosome_name" correctly defined?
#> ✔ Entries in chrom with the values "chr1" and "chr2" are retained.
#> ℹ The following 7 entries in column chrom from the input data were not
#>   retained: "Chr2", "chr10", "chr4 2", "chr4-2", "chr4|2", "chr42", and
#>   "chr4?2".
#> ✔ Input data was filtered to retain regions on defined chromosome.
#>   
#> → Blacklist for annotation "hg38" will be used for filtering.
#> ! Provided input data contains chromosome names not found in blacklist.
#> ℹ The following input chromosomes have no match: chr10, chr11, chr12, chr13,
#>   chr14, chr15, chr16, chr17, chr18, chr19, chr20, chr21, chr22, chr3, chr4,
#>   chr5, chr6, chr7, …, chrX, and chrY.
#> → Note to user: Please doublecheck this observation.
#> ✔ Input data was filtered by blacklist.
#>   
#> → Significance in score is filtered and all regions above 10 will be retained.
#> ℹ A total of 0 of 37 input regions are retained with value in score a above 10.
#> ✔ Input data was filtered to retain regions with a score above the defined
#>   threshold.
#>   
#> ℹ The argument `include_top_n_scoring` extracted the the top 100 regions by
#>   score per sample (based on the values in sample_name).
#> → The top enriched 100 regions per sample will be retained.
#> ✔ Input data was filtered and the top 100 enriched regions per sample are
#>   retained.
#>   
#> ✔ Filtered dataset will be returned.
#> # A tibble: 0 × 8
#> # ℹ 8 variables: chrom <chr>, start <int>, end <int>, name <chr>, score <dbl>,
#> #   strand <lgl>, center <dbl>, sample_name <chr>