Prepare input data for peakCombiner package

prepare_input_regions prepares the input data in the format needed for all of the following steps within peakCombiner. It accepts the following formats:

in memory data frame listing each sample's peak file location,

in memory data frame listing the peaks themselves that are found in each sample, or

in memory GRanges object listing the peaks themselves that are found in each sample.

prepare_input_regions(data, show_messages = TRUE)

Arguments

data

Data frame or GRanges object with the input data. Several formats are accepted, which are described in full in the Details below.

in memory data frame listing each sample's peak file location,

in memory data frame listing the peaks themselves that are found in each sample, or

in memory GRanges object listing the peaks themselves that are found in each sample.

show_messages

Logical value of TRUE (default) or FALSE. Defines if info messages are displayed or not.

Value

A tibble with the columns chrom, start, end, name, score, strand, center, sample_name. The definitions of these columns are described in full in the Details below. Use as input for functions center_expand_regions(), filter_regions() and combine_regions().

Details

Accepted inputs are one of the three following options:

In memory data frame listing each sample's peak file location

sample_name - Unique name for each sample (required).

path_peaks - Path to the file in which the genomic regions are stored. For example, the path to a bed file or .narrowPeak file (required).

file_format - The expected file format. Needed to correctly label the columns of the input. Acceptable values are: bed, narrowPeak, and broadPeak (required).

score_column - The number of the column having the the metric used to rank peak importance, where bigger values are more important. If not provided, column 9 will be used for .narrowPeak or .broadPeak file formats. Column 9 corresponds to the qValue as described in the UCSC documentation here. Other alternatives for narrowPeak or broadPeak could be columns 7 or 8, which correspond to signalValue or pValue (optional).

In memory data frame listing the peaks themselves that are found in each sample. The columns can be provided in any order and have the following names. Note that additional columns will be dropped.

chrom - chromosome name (required).

start - start coordinate of range (1-based coordinate system, NOT like bed files which are 0-based) (required).

end - end coordinate of range (required).

sample_name - unique identifier for a sample. No restrictions on characters (required).

score - the metric used to rank peak importance, where bigger values are more important. For example, qValue from Macs2, -log10FDR from another method, or fold enrichment over background computed from your favorite method. If not provided, defaults to 0 (optional).

strand - values are '+', '-', or '.'. If not provided, defaults to '.' (optional).

summit - distance of the strongest signal ("summit") of the peak region from the start coordinate (optional).

In memory GRanges object listing the peaks themselves that are found in each sample. This object is very similar to the data frame above, except that chrom, start, and end are instead described using the GRanges nomenclature. Note that additional columns will be dropped.

This function parses the inputs provided and returns a data frame having the columns listed below.

chrom - chromosome name

start - start coordinate of range (1-based coordinate system, NOT like bed files which are 0-based)

end - end coordinate of range

name - unique identifier for a region. auto-generated by this function

score - the metric used to rank peak importance, where bigger values are more important. For example, qValue from MACS2, -log10FDR from another method, or fold enrichment over background computed from your favorite method

strand - values are '+', '-', or '.'. Chromatin data are typically non- stranded so will have a '.'.

center - absolute genomic coordinate of the nucleotide at the center of the peak region, or alternatively the strongest signal ("summit") of the peak region. If no value is provided by the user, center defaults to the arithmetic center of the peak region.

sample_name - unique identifier for a sample. No restrictions on characters

In addition, input data is checked for mutliple entries of the same genomic region. This can occure when using called peak files as multiple summits can be annotated within the sqme genomic regions (defined by chrom, start and end). To avoid mutliple entries, this script is checking the input for multiple summits within the same regions and maintains only the strongest enriched (based on the values in the column score). This step is mantory to quaranty an optimal result.

Examples

# Load in and prepare a an accepted tibble
utils::data(syn_data_tibble)

data_prepared <- prepare_input_regions(
  data = syn_data_tibble,
  show_messages = TRUE
)
#> ℹ Provide input `data` is a pre-loaded <data.frame> with the required column
#>   names.
#> → Start preparing data.
#> ℹ Required columns will be added if missing.
#> ! Column 'name' from `data` will be overwritten.
#> → Column 'name' is a computed column from peakCombiner and therefore
#>   pre-exisiting data in a column 'name' will not be retained.
#> ! Column 'center' does exist in `data_prepared`.
#> → The column 'center' is taken to define output column 'center'.
#> → Checking <class> and "values" of all columns.
#> ✔ Structure of data was successfully checked to be an accepted input.
#>   
#> → Checking whether duplicated regions exist and need to be collapsed.
#> → Checked whether duplicated regions exist and need to be collapsed.
#> ✔ Duplicated regions identified and collapsed to unique chrom, start, and end
#>   for each sample by strongest score value.
#>   
#> ✔ Preparation of data finished successfully.
data_prepared
#> # A tibble: 52 × 8
#>    chrom  start   end name            score strand center sample_name 
#>    <chr>  <dbl> <dbl> <chr>           <dbl> <chr>   <dbl> <chr>       
#>  1 Chr2     101   800 control_rep1|9     80 .         700 control_rep1
#>  2 chr1     301  1000 control_rep1|1     96 .         600 control_rep1
#>  3 chr1    3501  4400 control_rep1|2     95 .        3800 control_rep1
#>  4 chr1    4701  5300 control_rep1|3     46 .        4900 control_rep1
#>  5 chr1    5601  6100 control_rep1|5     26 .        5700 control_rep1
#>  6 chr1    6701  7400 control_rep1|6     25 .        7000 control_rep1
#>  7 chr10    301  1000 control_rep1|7     75 .         600 control_rep1
#>  8 chr2     301  1000 control_rep1|8     50 .         600 control_rep1
#>  9 chr4 2   301  1000 control_rep1|10    30 .         600 control_rep1
#> 10 chr4-2   401  1100 control_rep1|11    20 .         600 control_rep1
#> # ℹ 42 more rows

# Or a pre-loaded tibble with genomic regions and named columns.

utils::data(syn_data_control01)
utils::data(syn_data_treatment01)

combined_input <- syn_data_control01 |>
  dplyr::mutate(sample_name = "control-rep1") |>
  rbind(syn_data_treatment01 |>
    dplyr::mutate(sample_name = "treatment-rep1"))

prepare_input_regions(
  data = combined_input,
  show_messages = FALSE
)
#> # A tibble: 19 × 8
#>    chrom  start   end name              score strand center sample_name   
#>    <chr>  <dbl> <dbl> <chr>             <dbl> <chr>   <dbl> <chr>         
#>  1 Chr2     101   800 control-rep1|9       80 .         700 control-rep1  
#>  2 chr1     301  1000 control-rep1|1       96 .         600 control-rep1  
#>  3 chr1    3501  4400 control-rep1|2       95 .        3800 control-rep1  
#>  4 chr1    4701  5300 control-rep1|3       46 .        4900 control-rep1  
#>  5 chr1    5601  6100 control-rep1|5       26 .        5700 control-rep1  
#>  6 chr1    6701  7400 control-rep1|6       25 .        7000 control-rep1  
#>  7 chr10    301  1000 control-rep1|7       75 .         600 control-rep1  
#>  8 chr2     301  1000 control-rep1|8       50 .         600 control-rep1  
#>  9 chr4 2   301  1000 control-rep1|10      30 .         600 control-rep1  
#> 10 chr4-2   401  1100 control-rep1|11      20 .         600 control-rep1  
#> 11 chr1     200   900 treatment-rep1|1    100 .         500 treatment-rep1
#> 12 chr1     201   900 treatment-rep1|2    100 .         500 treatment-rep1
#> 13 chr1    2601  3200 treatment-rep1|3     99 .        2800 treatment-rep1
#> 14 chr1    4501  5200 treatment-rep1|4     60 .        4700 treatment-rep1
#> 15 chr1    5801  6300 treatment-rep1|6     30 .        6100 treatment-rep1
#> 16 chr1    6701  7400 treatment-rep1|7     29 .        7000 treatment-rep1
#> 17 chr10    201   900 treatment-rep1|8     95 .         500 treatment-rep1
#> 18 chr2     201   900 treatment-rep1|9     50 .         500 treatment-rep1
#> 19 chr4-2   201   900 treatment-rep1|10    30 .         500 treatment-rep1